# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "abdxdev Portfolio API",
    "description": "A public REST API powering abdxdev's personal portfolio. Exposes endpoints for projects, skills, social links, friends, anime watchlist, game library, and visitor feedback.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment"
    }
  ],
  "paths": {
...
    "/api/portfolio/games": {
      "get": {
        "operationId": "getGamesList",
        "summary": "List game library",
        "description": "Returns the portfolio owner's game library from RAWG, covering all status categories: beaten, playing, on hold, plan to play, dropped, and owned.",
        "tags": [
          "Portfolio"
        ],
        "responses": {
          "200": {
            "description": "List of games",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            }
          },
          "404": {
            "description": "RAWG social not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Game": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "background_image": {
            "type": "string",
            "format": "uri"
          },
          "background_image_cropped": {
            "type": "string",
            "format": "uri",
            "description": "Cropped 600x400 image"
          },
          "rawg_link": {
            "type": "string",
            "format": "uri",
            "description": "RAWG page URL"
          },
          "released": {
            "type": "string",
            "description": "Release date"
          },
          "rating": {
            "type": "number",
            "description": "RAWG rating"
          },
          "status": {
            "type": "string",
            "enum": [
              "beaten",
              "playing",
              "yet",
              "toplay",
              "dropped",
              "owned"
            ]
          }
        }
      },
...
    }
  }
}